requirements.txt
streamlit
pandas
numpy
requests
SQLAlchemy
psycopg2-binary
scikit-learn
plotly
python-dotenv
